+++ src/lib.rs
@@ -43,7 +43,7 @@
 //! There are many kind of events, see the documentation to know more.
 //!
 //! Through a `server` or a `client` you'll be able to access the full state of the battle,
-//! including the full timeline of events.
+//! including the entire timeline of events.
 //!
 //! ## Define the game's rules via traits
 //!